requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0

# Analysis dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Web Application (Dash)
dash>=2.14.0
dash-bootstrap-components>=1.5.0
